- Create dist folder containing modules transpiled into commonjs compatible ES5 - Updated NPM Build process to expose dist for js, sass, images, nunjucks - Moved gallery starter file and config into own folder so not exported - Updated package description to point to exported javascript
