+++ .babelrc
@@ -1,4 +1,4 @@
 {
   "presets": ["es2015"],
-  "plugins": ["transform-object-rest-spread", "transform-class-properties"]
+  "plugins": ["transform-object-rest-spread", "transform-class-properties", "babel-plugin-add-module-exports"]
 }
