+++ CHANGELOG.md
@@ -15,3 +15,10 @@ Move project to new repo.
 - Fix issue with add another when using controls other than textbox
 - Expanded gallery to include tabs an add more examples.
 - Improve gallery page layout.
+
+# 0.4.0
+
+- Create dist folder containing modules transpiled into commonjs compatible ES5
+- Updated NPM Build process to expose dist for js, sass, images, nunjucks
+- Moved gallery starter file and config into own folder so not exported
+- Updated package description to point to exported javascript
