+++ gallery/config.js
