+++ gulp/tasks/babel.js
@@ -1,0 +1,10 @@
+const gulp = require('gulp');
+const babel = require('gulp-babel');
+const paths = require('../paths');
+
+
+gulp.task('babel', () => {
+  return gulp.src(paths.libJS + '/*.js')
+    .pipe(babel())
+    .pipe(gulp.dest('dist'));
+});
