+++ gulp/tasks/watch.js
@@ -5,8 +5,6 @@ const gulpSequence = require('gulp-sequence');
 gulp.task('watch', (done) => {
 
   gulpSequence('build', 'serve', 'browserSync', () => {
-    const browserSync = require('browser-sync').get('bs-proxy');
-console.log('libjs: ' + paths.libJS);
     gulp.watch(`${paths.sourceJS}/**/*.js`, ['webpack']);
     gulp.watch(`${paths.libJS}/**/*.js`, ['webpack']);
     gulp.watch(`${paths.sourceStyles}/**/*.scss`, ['css']);
