+++ javascripts/radiohide.js
@@ -23,7 +23,7 @@ class RadioHideComponent {
         }
       }
     });
-  }
+  };
 
 }
