+++ sass/_horizontal-tabs.scss
