+++ sass/_results-list.scss
