+++ sass/_table-detail.scss
