+++ sass/_table.scss
